Refactor addon apps to not import models
