+++ addons/base/apps.py
@@ -8,6 +8,7 @@ from mako.lookup import TemplateLookup
 from framework.routing import process_rules
 from framework.flask import app
 from website import settings
+from website.util import rubeus
 
 
 def _is_image(filename):
@@ -26,6 +27,26 @@ USER_SETTINGS_TEMPLATE_DEFAULT = os.path.join(
     'profile',
     'user_settings_default.mako',
 )
+
+
+def generic_root_folder(addon_short_name):
+    def _root_folder(node_settings, auth, **kwargs):
+        """Return the Rubeus/HGrid-formatted response for the root folder only."""
+        # Quit if node settings does not have authentication
+        if not node_settings.has_auth or not node_settings.folder_id:
+            return None
+        node = node_settings.owner
+        root = rubeus.build_addon_root(
+            node_settings=node_settings,
+            name=node_settings.fetch_folder_name(),
+            permissions=auth,
+            nodeUrl=node.url,
+            nodeApiUrl=node.api_url,
+            private_key=kwargs.get('view_only', None),
+        )
+        return [root]
+    _root_folder.__name__ = '{0}_root_folder'.format(addon_short_name)
+    return _root_folder
 
 
 class BaseAddonAppConfig(AppConfig):
