+++ addons/dropbox/apps.py
@@ -1,8 +1,7 @@
-from addons.base.apps import BaseAddonAppConfig
+from addons.base.apps import BaseAddonAppConfig, generic_root_folder
 
-from addons.dropbox.views import dropbox_root_folder
-from addons.dropbox import routes
 
+dropbox_root_folder = generic_root_folder('dropbox')
 
 class DropboxAddonAppConfig(BaseAddonAppConfig):
 
@@ -13,13 +12,17 @@ class DropboxAddonAppConfig(BaseAddonAppConfig):
     configs = ['accounts', 'node']
     has_hgrid_files = True
     max_file_size = 150  # MB
-    routes = [routes.auth_routes, routes.api_routes]
     owners = ['user', 'node']
     categories = ['storage']
 
     @property
     def get_hgrid_data(self):
         return dropbox_root_folder
+
+    @property
+    def routes(self):
+        from . import routes
+        return [routes.auth_routes, routes.api_routes]
 
     FOLDER_SELECTED = 'dropbox_folder_selected'
     NODE_AUTHORIZED = 'dropbox_node_authorized'
