+++ addons/owncloud/views.py
@@ -35,10 +35,6 @@ owncloud_deauthorize_node = generic_views.deauthorize_node(
     SHORT_NAME
 )
 
-owncloud_root_folder = generic_views.root_folder(
-    SHORT_NAME
-)
-
 ## Config ##
 
 @must_be_logged_in
